Switch to an XML based format for bundle declarations
This replaces the line-based format with an XML format. Profiling shows
that, once the JIT is warmed up, the average XML declaration file
takes around 16µs to parse. The old line-based format was not much
faster than this. Additionally, the file is parsed with a stream parser
and therefore does not require allocating any large AST structures.
The new XML format has a well-defined XSD schema and allows working
with standard tools for up-front validation, error reporting, etc.