Redesign string table handling
This redesigns the schema and parser so that the user is required to
provide a default text string and a set of optional translations. If
the user requests a string table with a language that is not supported
by that table, the user will get a table with the default strings
instead.