Rename receipts, implement ping/pong
This renames receipts to acks, and avoids sending them unless there's
something that actually requires a negative ACK. This also required
implementing ping/pong packets in order to avoid having connections
time out due to a lack of packets to send/receive. In the process,
this also augments connections with RTT information, although the
implementation imposes a lower bound for this information: A 30hz
tick rate means that a receiver can't respond with a pong packet in
less than 15ms, so the lower bound becomes 2 * 15ms = 30ms.

Fix #1